// windmill/f/actual_up/types.ts
export type ActualUpConfig = {
  actual: {
    bridgeToken: string;
    bridgeUrl: string;
  };
  alerts?: { url: string };
  categoryMappings: Record<string, string>;
  mappings: AccountMapping[];
  notes: {
    includeCashback: boolean;
    includeForeignAmount: boolean;
    includeMessage: boolean;
    includePerformer: boolean;
    includeRoundUp: boolean;
  };
  schedule: { lookbackDays: number; timezone: string };
  up: { connections: UpConnection[] };
};

export type AccountMapping = {
  actualAccountId: string;
  alias: string;
  connections: string[];
  upAccountId: string;
};

export type UpConnection = {
  id: string;
  token: string;
  webhook?: { id: string; secret: string };
};